# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Course System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Swagger-doc Course System API"
        }
    ],
    "paths": {
...
        "/api/course/create": {
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "建立新課程",
                "description": "暫無描述。",
                "operationId": "395cc55f6f29c0fdeec6369e2b0c9029",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "num": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "credits": {
                                        "type": "integer"
                                    },
                                    "limit_people": {
                                        "type": "integer"
                                    },
                                    "classroom_and_time": {
                                        "type": "string"
                                    },
                                    "instructors": {
                                        "type": "string"
                                    },
                                    "limit_illustrate": {
                                        "type": "string"
                                    },
                                    "required_illustrate": {
                                        "type": "string"
                                    },
                                    "remark": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "num": "2301A051",
                                    "name": "資料結構",
                                    "credits": 3,
                                    "limit_people": 40,
                                    "classroom_and_time": "管理學院 資訊大樓303，週一 13:10",
                                    "instructors": "P0002",
                                    "limit_illustrate": "限資管系",
                                    "required_illustrate": "資訊管理系必修",
                                    "remark": "無"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": {
                                            "id": 7,
                                            "num": "2301A051",
                                            "name": "資料結構",
                                            "credits": 3,
                                            "limit_people": 40,
                                            "classroom_and_time": "管理學院 資訊大樓303，週一 13:10",
                                            "instructors": "黃電子",
                                            "limit_illustrate": "限資管系",
                                            "required_illustrate": "資訊管理系必修",
                                            "remark": "無"
                                        },
                                        "TimeStamp": "2023-09-07T03:34:32.580848+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/course/show/{id}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "取得課程單一資訊",
                "description": "暫無描述。",
                "operationId": "c8e83e2fab39b9ac5b17c7beb693d04f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": {
                                            "id": 3,
                                            "num": "2301B001",
                                            "name": "電子學",
                                            "credits": 3,
                                            "limit_people": 40,
                                            "classroom_and_time": "工程學院 電子系大樓201，週二 13:10",
                                            "instructors": "黃電子",
                                            "limit_illustrate": "限工程學院所有科系",
                                            "required_illustrate": "電子系必修",
                                            "remark": "本課程為18週課程"
                                        },
                                        "TimeStamp": "2023-09-07T03:39:26.390217+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/course/update/{id}": {
            "put": {
                "tags": [
                    "Course"
                ],
                "summary": "更新課程內容",
                "description": "暫無描述。",
                "operationId": "99b07001e06e61eed6834c3922cb817a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "4"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "credits": {
                                        "type": "integer"
                                    },
                                    "limit_people": {
                                        "type": "integer"
                                    },
                                    "classroom_and_time": {
                                        "type": "string"
                                    },
                                    "instructors": {
                                        "type": "string"
                                    },
                                    "limit_illustrate": {
                                        "type": "string"
                                    },
                                    "required_illustrate": {
                                        "type": "string"
                                    },
                                    "remark": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "數位多媒體設計",
                                    "credits": 3,
                                    "limit_people": 60,
                                    "classroom_and_time": "創意媒體學院1館701，週四 13:10",
                                    "instructors": "P0003",
                                    "limit_illustrate": "限多媒體學院",
                                    "required_illustrate": "資訊傳播系必修",
                                    "remark": "本課程需做成果展示"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": {
                                            "id": 4,
                                            "num": "2301C002",
                                            "name": "數位多媒體設計",
                                            "credits": 3,
                                            "limit_people": 60,
                                            "classroom_and_time": "創意媒體學院1館701，週四 13:10",
                                            "instructors": "陳媒體",
                                            "limit_illustrate": "限多媒體學院",
                                            "required_illustrate": "資訊傳播系必修",
                                            "remark": "本課程需做成果展示"
                                        },
                                        "TimeStamp": "2023-09-07T03:43:25.323960+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/course/delete/{id}": {
            "delete": {
                "tags": [
                    "Course"
                ],
                "summary": "刪除課程",
                "description": "暫無描述。",
                "operationId": "09c05029c5d57bc591edce1d52589ed2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": "體育課程已刪除。",
                                        "TimeStamp": "2023-09-07T03:44:56.903975+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}